stock À۾÷ ÁøÇàÁß
