--- src/main/java/com/example/freshcart/optionstock/domain/JpaProductStockRepository.java
@@ -1,10 +1,0 @@
-package com.example.freshcart.optionstock.domain;
-
-import java.util.Optional;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface JpaProductStockRepository extends JpaRepository<ProductStock, Long> {
-
-  ProductStock findByProductId(Long productId);
-  Optional<ProductStock> findById(Long productStockId);
-}
