+++ src/main/java/com/example/freshcart/optionstock/domain/ProductStockRepository.java
@@ -4,6 +4,8 @@ import com.example.freshcart.product.domain.Product;
 
 public interface ProductStockRepository {
   void save(ProductStock productStock);
+  ProductStock findByProductIdWithPessimisticLock(Long productId);
+
   ProductStock findByProductId(Long productId);
   ProductStock findById(Long productStockId);
