+++ src/main/java/com/example/freshcart/order/infrastructure/config/OrderRepositoryConfig.java
@@ -1,21 +1,14 @@
 package com.example.freshcart.order.infrastructure.config;
 
-import com.example.freshcart.optionstock.domain.JpaOptionStockRepository;
 import com.example.freshcart.order.domain.JpaOrderItemOptionRepository;
 import com.example.freshcart.order.domain.JpaOrderItemRepository;
 import com.example.freshcart.order.domain.JpaOrderRepository;
 import com.example.freshcart.order.domain.OrderItemOptionRepository;
 import com.example.freshcart.order.domain.OrderItemRepository;
 import com.example.freshcart.order.domain.OrderRepository;
-import com.example.freshcart.order.infrastructure.OrderItemOptionMapperAdaptor;
-import com.example.freshcart.order.infrastructure.OrderItemMapperAdaptor;
 import com.example.freshcart.order.infrastructure.jpa.JpaOrderItemOptionRepositoryAdaptor;
 import com.example.freshcart.order.infrastructure.jpa.JpaOrderItemRepositoryAdaptor;
 import com.example.freshcart.order.infrastructure.jpa.JpaOrderRepositoryAdaptor;
-import com.example.freshcart.order.infrastructure.mybatis.OrderItemMapper;
-import com.example.freshcart.order.infrastructure.mybatis.OrderItemOptionMapper;
-import com.example.freshcart.order.infrastructure.mybatis.OrderMapper;
-import com.example.freshcart.order.infrastructure.OrderMapperAdaptor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
